requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
Faker==38.2.0
fhir.resources==8.1.0
fhir_core==1.1.4
ftfy==6.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
postgrest==2.24.0
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.24.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
